+++ tests/testthat/test-translate-sql.R
@@ -109,6 +109,11 @@ test_that("if and ifelse use correctly named arguments",{
 
   expect_equal(translate_sql(ifelse(test = x, yes = 1, no = 2)), exp)
   expect_equal(translate_sql(if_else(condition = x, true = 1, false = 2)), exp)
+
+  expect_equal(
+    translate_sql(if_else(condition = x, true = 1, false = 2, missing = 3)),
+    sql("CASE WHEN (`x`) THEN (1.0) WHEN NOT(`x`) THEN (2.0) WHEN ((`x`) IS NULL) THEN (3.0) END")
+  )
 })
 
 test_that("switch translated to CASE WHEN", {